add condition statements to check if connected to DB

+++ src/Service/ServiceArticle.java
@@ -33,6 +33,10 @@ public class ServiceArticle implements IServiceArticle {
     
     @Override
     public void addArticle(Article a) {
+         if(cnx == null){
+             Utilities.alert("ERROR", "could not connect to database.");
+            return;
+        }
         try {
             Statement stm = cnx.createStatement();
             System.out.println("of: " +String.valueOf(a.getArticle_id()) 
@@ -60,6 +64,10 @@ public class ServiceArticle implements IServiceArticle {
 
     @Override
     public List<Article> showArticle() {
+        if(cnx == null){
+            Utilities.alert("ERROR", "could not connect to database.");
+            return null;
+        }
         List<Article> list = new ArrayList<>();
         try{
             Statement stm = this.cnx.createStatement();
@@ -81,12 +89,17 @@ public class ServiceArticle implements IServiceArticle {
             }
         }catch(SQLException ex){
             System.out.println("Could not show articles");
+            list = null;
         }
         return list;
     }
 
     @Override
     public void modArticle(Article a) {
+         if(cnx == null){
+             Utilities.alert("ERROR", "could not connect to database.");
+            return;
+        }
         try{
             Statement stm = cnx.createStatement();
             String query = "update article set "
@@ -105,12 +118,16 @@ public class ServiceArticle implements IServiceArticle {
 
     @Override
     public void delArticle(Article a) {
+         if(cnx == null){
+             Utilities.alert("ERROR", "could not connect to database.");
+            return;
+        }
         try{
             Statement stm = cnx.createStatement();
             String query = "delete from article where article_id="+a.getArticle_id()+";";
             stm.executeUpdate(query);
         }catch(SQLException ex){
-            System.out.println("Could not modify article");
+            System.out.println("Could not delete article");
         }
     }
